+++ sdk/ai/azure-ai-projects/samples/evaluations/agentic_evaluators/sample_groundedness.py
@@ -36,6 +36,7 @@ from openai.types.evals.create_eval_jsonl_run_data_source_param import (
     SourceFileContent,
     SourceFileContentContent,
 )
+from openai.types.eval_create_params import DataSourceConfigCustom
 
 
 load_dotenv()
@@ -53,26 +54,28 @@ def main() -> None:
 
             client = project_client.get_openai_client()
 
-            data_source_config = {
-                "type": "custom",
-                "item_schema": {
-                    "type": "object",
-                    "properties": {
-                        "context": {"type": "string"},
-                        "query": {"anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "object"}}]},
-                        "response": {"anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "object"}}]},
-                        "tool_definitions": {
-                            "anyOf": [
-                                {"type": "string"},
-                                {"type": "object"},
-                                {"type": "array", "items": {"type": "object"}},
-                            ]
+            data_source_config = DataSourceConfigCustom(
+                {
+                    "type": "custom",
+                    "item_schema": {
+                        "type": "object",
+                        "properties": {
+                            "context": {"type": "string"},
+                            "query": {"anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "object"}}]},
+                            "response": {"anyOf": [{"type": "string"}, {"type": "array", "items": {"type": "object"}}]},
+                            "tool_definitions": {
+                                "anyOf": [
+                                    {"type": "string"},
+                                    {"type": "object"},
+                                    {"type": "array", "items": {"type": "object"}},
+                                ]
+                            },
                         },
+                        "required": ["response"],
                     },
-                    "required": ["response"],
-                },
-                "include_sample_schema": True,
-            }
+                    "include_sample_schema": True,
+                }
+            )
 
             testing_criteria = [
                 {
@@ -92,8 +95,8 @@ def main() -> None:
             print("Creating Eval Group")
             eval_object = client.evals.create(
                 name="Test Groundedness Evaluator with inline data",
-                data_source_config=data_source_config, # type: ignore
-                testing_criteria=testing_criteria, # type: ignore
+                data_source_config=data_source_config,
+                testing_criteria=testing_criteria,  # type: ignore
             )
             print(f"Eval Group created")