+++ sdk/ai/azure-ai-projects/samples/evaluations/sample_model_evaluation.py
@@ -28,11 +28,13 @@ USAGE:
 import os
 import time
 from pprint import pprint
+from typing import Union
 from dotenv import load_dotenv
 from azure.identity import DefaultAzureCredential
 from azure.ai.projects import AIProjectClient
-from azure.ai.projects.models import PromptAgentDefinition
 from openai.types.eval_create_params import DataSourceConfigCustom
+from openai.types.evals.run_create_response import RunCreateResponse
+from openai.types.evals.run_retrieve_response import RunRetrieveResponse
 
 load_dotenv()
 
@@ -60,8 +62,8 @@ with project_client:
     ]
     eval_object = openai_client.evals.create(
         name="Agent Evaluation",
-        data_source_config=data_source_config, # type: ignore
-        testing_criteria=testing_criteria, # type: ignore
+        data_source_config=data_source_config,
+        testing_criteria=testing_criteria,  # type: ignore
     )
     print(f"Evaluation created (id: {eval_object.id}, name: {eval_object.name})")
 
@@ -91,8 +93,8 @@ with project_client:
         },
     }
 
-    agent_eval_run = openai_client.evals.runs.create(
-        eval_id=eval_object.id, name=f"Evaluation Run for Model {model}", data_source=data_source
+    agent_eval_run: Union[RunCreateResponse, RunRetrieveResponse] = openai_client.evals.runs.create(
+        eval_id=eval_object.id, name=f"Evaluation Run for Model {model}", data_source=data_source  # type: ignore
     )
     print(f"Evaluation run created (id: {agent_eval_run.id})")