+++ sdk/paloaltonetworks/azure-mgmt-paloaltonetworksngfw/azure/mgmt/paloaltonetworksngfw/aio/_palo_alto_networks_ngfw_mgmt_client.py
@@ -7,13 +7,18 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, Awaitable, TYPE_CHECKING
+from typing import Any, Awaitable, Optional, TYPE_CHECKING, cast
+from typing_extensions import Self
 
+from azure.core.pipeline import policies
 from azure.core.rest import AsyncHttpResponse, HttpRequest
+from azure.core.settings import settings
 from azure.mgmt.core import AsyncARMPipelineClient
+from azure.mgmt.core.policies import AsyncARMAutoResourceProviderRegistrationPolicy
+from azure.mgmt.core.tools import get_arm_endpoints
 
 from .. import models as _models
-from .._serialization import Deserializer, Serializer
+from .._utils.serialization import Deserializer, Serializer
 from ._configuration import PaloAltoNetworksNgfwMgmtClientConfiguration
 from .operations import (
     CertificateObjectGlobalRulestackOperations,
@@ -25,7 +30,9 @@ from .operations import (
     GlobalRulestackOperations,
     LocalRulesOperations,
     LocalRulestacksOperations,
+    MetricsObjectFirewallOperations,
     Operations,
+    PaloAltoNetworksCloudngfwOperationsOperations,
     PostRulesOperations,
     PreRulesOperations,
     PrefixListGlobalRulestackOperations,
@@ -33,11 +40,11 @@ from .operations import (
 )
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
+    from azure.core import AzureClouds
     from azure.core.credentials_async import AsyncTokenCredential
 
 
-class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-version-keyword,too-many-instance-attributes
+class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=too-many-instance-attributes
     """PaloAltoNetworksNgfwMgmtClient.
 
     :ivar global_rulestack: GlobalRulestackOperations operations
@@ -52,18 +59,25 @@ class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-vers
      azure.mgmt.paloaltonetworksngfw.aio.operations.FqdnListGlobalRulestackOperations
     :ivar post_rules: PostRulesOperations operations
     :vartype post_rules: azure.mgmt.paloaltonetworksngfw.aio.operations.PostRulesOperations
+    :ivar pre_rules: PreRulesOperations operations
+    :vartype pre_rules: azure.mgmt.paloaltonetworksngfw.aio.operations.PreRulesOperations
     :ivar prefix_list_global_rulestack: PrefixListGlobalRulestackOperations operations
     :vartype prefix_list_global_rulestack:
      azure.mgmt.paloaltonetworksngfw.aio.operations.PrefixListGlobalRulestackOperations
-    :ivar pre_rules: PreRulesOperations operations
-    :vartype pre_rules: azure.mgmt.paloaltonetworksngfw.aio.operations.PreRulesOperations
     :ivar operations: Operations operations
     :vartype operations: azure.mgmt.paloaltonetworksngfw.aio.operations.Operations
+    :ivar palo_alto_networks_cloudngfw_operations: PaloAltoNetworksCloudngfwOperationsOperations
+     operations
+    :vartype palo_alto_networks_cloudngfw_operations:
+     azure.mgmt.paloaltonetworksngfw.aio.operations.PaloAltoNetworksCloudngfwOperationsOperations
     :ivar firewalls: FirewallsOperations operations
     :vartype firewalls: azure.mgmt.paloaltonetworksngfw.aio.operations.FirewallsOperations
     :ivar local_rulestacks: LocalRulestacksOperations operations
     :vartype local_rulestacks:
      azure.mgmt.paloaltonetworksngfw.aio.operations.LocalRulestacksOperations
+    :ivar metrics_object_firewall: MetricsObjectFirewallOperations operations
+    :vartype metrics_object_firewall:
+     azure.mgmt.paloaltonetworksngfw.aio.operations.MetricsObjectFirewallOperations
     :ivar firewall_status: FirewallStatusOperations operations
     :vartype firewall_status:
      azure.mgmt.paloaltonetworksngfw.aio.operations.FirewallStatusOperations
@@ -82,9 +96,12 @@ class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-vers
     :type credential: ~azure.core.credentials_async.AsyncTokenCredential
     :param subscription_id: The ID of the target subscription. Required.
     :type subscription_id: str
-    :param base_url: Service URL. Default value is "https://management.azure.com".
+    :param base_url: Service URL. Default value is None.
     :type base_url: str
-    :keyword api_version: Api Version. Default value is "2023-09-01". Note that overriding this
+    :keyword cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
+     None.
+    :paramtype cloud_setting: ~azure.core.AzureClouds
+    :keyword api_version: Api Version. Default value is "2025-10-08". Note that overriding this
      default value may result in unsupported behavior.
     :paramtype api_version: str
     :keyword int polling_interval: Default waiting time between two polls for LRO operations if no
@@ -95,13 +112,45 @@ class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-vers
         self,
         credential: "AsyncTokenCredential",
         subscription_id: str,
-        base_url: str = "https://management.azure.com",
+        base_url: Optional[str] = None,
+        *,
+        cloud_setting: Optional["AzureClouds"] = None,
         **kwargs: Any
     ) -> None:
+        _cloud = cloud_setting or settings.current.azure_cloud  # type: ignore
+        _endpoints = get_arm_endpoints(_cloud)
+        if not base_url:
+            base_url = _endpoints["resource_manager"]
+        credential_scopes = kwargs.pop("credential_scopes", _endpoints["credential_scopes"])
         self._config = PaloAltoNetworksNgfwMgmtClientConfiguration(
-            credential=credential, subscription_id=subscription_id, **kwargs
+            credential=credential,
+            subscription_id=subscription_id,
+            cloud_setting=cloud_setting,
+            credential_scopes=credential_scopes,
+            **kwargs
         )
-        self._client: AsyncARMPipelineClient = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+
+        _policies = kwargs.pop("policies", None)
+        if _policies is None:
+            _policies = [
+                policies.RequestIdPolicy(**kwargs),
+                self._config.headers_policy,
+                self._config.user_agent_policy,
+                self._config.proxy_policy,
+                policies.ContentDecodePolicy(**kwargs),
+                AsyncARMAutoResourceProviderRegistrationPolicy(),
+                self._config.redirect_policy,
+                self._config.retry_policy,
+                self._config.authentication_policy,
+                self._config.custom_hook_policy,
+                self._config.logging_policy,
+                policies.DistributedTracingPolicy(**kwargs),
+                policies.SensitiveHeaderCleanupPolicy(**kwargs) if self._config.redirect_policy else None,
+                self._config.http_logging_policy,
+            ]
+        self._client: AsyncARMPipelineClient = AsyncARMPipelineClient(
+            base_url=cast(str, base_url), policies=_policies, **kwargs
+        )
 
         client_models = {k: v for k, v in _models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
@@ -117,13 +166,19 @@ class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-vers
             self._client, self._config, self._serialize, self._deserialize
         )
         self.post_rules = PostRulesOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.pre_rules = PreRulesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.prefix_list_global_rulestack = PrefixListGlobalRulestackOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
-        self.pre_rules = PreRulesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.operations = Operations(self._client, self._config, self._serialize, self._deserialize)
+        self.palo_alto_networks_cloudngfw_operations = PaloAltoNetworksCloudngfwOperationsOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
         self.firewalls = FirewallsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.local_rulestacks = LocalRulestacksOperations(
+            self._client, self._config, self._serialize, self._deserialize
+        )
+        self.metrics_object_firewall = MetricsObjectFirewallOperations(
             self._client, self._config, self._serialize, self._deserialize
         )
         self.firewall_status = FirewallStatusOperations(self._client, self._config, self._serialize, self._deserialize)
@@ -138,7 +193,9 @@ class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-vers
             self._client, self._config, self._serialize, self._deserialize
         )
 
-    def _send_request(self, request: HttpRequest, **kwargs: Any) -> Awaitable[AsyncHttpResponse]:
+    def _send_request(
+        self, request: HttpRequest, *, stream: bool = False, **kwargs: Any
+    ) -> Awaitable[AsyncHttpResponse]:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
@@ -158,12 +215,12 @@ class PaloAltoNetworksNgfwMgmtClient:  # pylint: disable=client-accepts-api-vers
 
         request_copy = deepcopy(request)
         request_copy.url = self._client.format_url(request_copy.url)
-        return self._client.send_request(request_copy, **kwargs)
+        return self._client.send_request(request_copy, stream=stream, **kwargs)  # type: ignore
 
     async def close(self) -> None:
         await self._client.close()
 
-    async def __aenter__(self) -> "PaloAltoNetworksNgfwMgmtClient":
+    async def __aenter__(self) -> Self:
         await self._client.__aenter__()
         return self