+++ sdk/ai/azure-ai-projects/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 2.0.0b2 (Unreleased)
+
+### Features Added
+
+### Breaking changes
+
+* `get_openai_client()` method on the asynchronous AIProjectClient is no longer an "async" method.
+* tracing: tool call output event content format updated to be in line with other events
+
+### Bugs Fixed
+
+### Sample updates
+* Added `finetuning` samples for operations create, retrieve, list, list_events, list_checkpoints, cancel, pause and resume. Also, these samples includes various finetuning techniques like Supervised (SFT), Reinforcement (RFT) and Direct performance optimization (DPO).
+
 ## 2.0.0b1 (2025-11-11)
 
 ### Features added